requirements.txt
gspread==3.7.0
python-dotenv==0.15.0
requests==2.25.1
